Profile creator Ui first steps
Added bootstrap, query and other resources for building the skeleton
(not completed) of the profile creator. The idea is that users can
either provide the url where their profile info is stored, which would
be loaded in the form to edit. Or start the form from scratch. Once the
user is done and presses the button, the chunk of text containing the
rdf-formatted data will be outputted for download or copy/paste.